Refactor into a sub-package.
There are going to be several persistent classes; they will be easier
to maintain if they are in separate files.
